+++ src/image-alt-text/auditEngine.js
@@ -12,6 +12,7 @@
 
 import { isNonEmptyArray, hasText } from '@adobe/spacecat-shared-utils';
 import { Audit as AuditModel } from '@adobe/spacecat-shared-data-access';
+import { franc } from 'franc-min';
 
 // GPT support: https://platform.openai.com/docs/guides/vision
 const SUPPORTED_FORMATS = /\.(webp|png|gif|jpeg|jpg)(?=\?|$)/i;
@@ -23,6 +24,7 @@ const mimeTypesForBase64 = {
   ico: 'image/x-icon',
 };
 const AUDIT_TYPE = AuditModel.AUDIT_TYPES.ALT_TEXT;
+const UNKNOWN_LANGUAGE = 'unknown';
 
 const getMimeType = async (url) => {
   const match = url.match(SUPPORTED_BLOB_FORMATS);
@@ -70,36 +72,78 @@ export const convertImagesToBase64 = async (imageUrls, auditUrl, log, fetch) => 
   return base64Blobs;
 };
 
+function detectLanguageFromText(text) {
+  const langCode = franc(text);
+  return langCode !== 'und' ? langCode : UNKNOWN_LANGUAGE;
+}
+
+function detectLanguageFromDom({ document }) {
+  const htmlTag = document.querySelector('html');
+  if (htmlTag && htmlTag.hasAttribute('lang')) {
+    return htmlTag.getAttribute('lang');
+  }
+
+  const metaTags = document.querySelectorAll('meta[http-equiv="Content-Language"], meta[name="language"]');
+  for (const meta of metaTags) {
+    if (meta.hasAttribute('content')) {
+      return meta.getAttribute('content');
+    }
+  }
+
+  return UNKNOWN_LANGUAGE;
+}
+
+const getPageLanguage = ({ document }) => {
+  let lang = UNKNOWN_LANGUAGE;
+  if (!document) {
+    return lang;
+  }
+
+  lang = detectLanguageFromDom({ document });
+  if (lang === UNKNOWN_LANGUAGE) {
+    const bodyText = document.querySelector('body').textContent;
+    lang = detectLanguageFromText(bodyText);
+  }
+  return lang;
+};
+
 export default class AuditEngine {
   constructor(log) {
     this.log = log;
     this.auditedTags = {
       imagesWithoutAltText: new Map(),
+      presentationalImagesWithoutAltText: new Map(),
     };
   }
 
   performPageAudit(pageUrl, pageTags) {
-    const presentationalImages = [];
     if (!isNonEmptyArray(pageTags?.images)) {
       this.log.debug(`[${AUDIT_TYPE}]: No images found for page ${pageUrl}`);
       return;
     }
 
-    pageTags.images.forEach((image) => {
-      if (image.isPresentational) {
-        presentationalImages.push(image);
-        return;
-      }
+    const pageLanguage = getPageLanguage({ document: pageTags.dom?.window?.document });
 
+    this.log.debug(`[${AUDIT_TYPE}]: Language: ${pageLanguage}, Page: ${pageUrl}`);
+
+    pageTags.images.forEach((image) => {
       if (!hasText(image.alt?.trim())) {
+        if (image.isPresentational) {
+          this.auditedTags.presentationalImagesWithoutAltText.set(image.src, {
+            pageUrl,
+            src: image.src,
+            xpath: image.xpath,
+          });
+        }
+
         this.auditedTags.imagesWithoutAltText.set(image.src, {
           pageUrl,
           src: image.src,
+          xpath: image.xpath,
+          language: pageLanguage,
         });
       }
     });
-
-    this.log.info(`[${AUDIT_TYPE}]: Presentational images:`, presentationalImages.length);
   }
 
   async filterImages(baseURL, fetch) {
@@ -139,6 +183,8 @@ export default class AuditEngine {
           blob: !!data.blob,
         })),
       );
+      // Log total blobs
+      this.log.info(`[${AUDIT_TYPE}]: Total blobs:`, base64Blobs.length);
 
       // Add unique blobs to the filtered map
       uniqueBlobsMap.forEach((originalData) => {
@@ -156,11 +202,19 @@ export default class AuditEngine {
     this.log.info(
       `[${AUDIT_TYPE}]: Found ${Array.from(this.auditedTags.imagesWithoutAltText.values()).length} images without alt text`,
     );
+    this.log.info(
+      `[${AUDIT_TYPE}]: Found ${Array.from(this.auditedTags.presentationalImagesWithoutAltText.values()).length} presentational images`,
+    );
   }
 
   getAuditedTags() {
     return {
       imagesWithoutAltText: Array.from(this.auditedTags.imagesWithoutAltText.values()),
+      presentationalImagesCount: Array.from(
+        this.auditedTags.presentationalImagesWithoutAltText.values(),
+      ).length,
     };
   }
 }
+
+export { getPageLanguage, detectLanguageFromText };