+++ src/index.js
@@ -37,6 +37,7 @@ import metaTags from './metatags/handler.js';
 import costs from './costs/handler.js';
 import structuredData from './structured-data/handler.js';
 import siteDetection from './site-detection/handler.js';
+import highFormViewsLowConversionsGuidance from './forms-opportunities/oppty-handlers/guidance-high-form-views-low-conversions.js';
 import highOrganicLowCtrGuidance from './experimentation-opportunities/guidance-high-organic-low-ctr-handler.js';
 import imageAltText from './image-alt-text/handler.js';
 
@@ -62,6 +63,7 @@ const HANDLERS = {
   'site-detection': siteDetection,
   'guidance:high-organic-low-ctr': highOrganicLowCtrGuidance,
   'alt-text': imageAltText,
+  'guidance:high-form-views-low-conversions': highFormViewsLowConversionsGuidance,
   dummy: (message) => ok(message),
 };
 
@@ -82,6 +84,7 @@ async function run(message, context) {
   const { type, siteId } = message;
 
   log.info(`Received ${type} audit request for: ${siteId}`);
+  log.info(`Message ${message}`);
 
   const handler = HANDLERS[type];
   if (!handler) {